Change single() to die rather than return undef
Two fo the four official Neo4j drivers offer a single() method. One of these (the Java driver) throws an exception if there is not exactly one record to be returned, the other one (for Python) simply returns the first record (and issues a warning). Given that the advantages of these two options aren't immediately obvious and this Perl driver is completely new, it seems sensible to start out with the more restrictive approach.

Should it turn out later that returning the first record is in fact the preferable behaviour for single(), it's easy to change the single() method to allow that. Allowing it now and changing it to disallow later would be much more difficult.

The existing behaviour of returning undef rather than throwing an exception hides the origin of the problem and was therefore a poor design choice.