Implement ResultSummary
The summary is not quite complete. In particular, the plan profile is missing, as well as the server timing information. I'm not sure either of these is provided via HTTP.

There is no has_plan() method, as there is no advantage over calling plan() directly and evaluating the result as boolean.