Track whether transactions are open or closed
Transactions are closed by the server on commit or rollback, and the server even informs us of that. This change tracks the open/closed status and fails immediately if any operations are attempted on closed transactions. This mirrors the behaviour of the official Neo4j drivers.

The server also reports transaction timeout durations. This software doesn't track those, as they are generally 60 seconds and it is believed that clients of this software wil complete transactions far sooner than that. Parsing timestamps over and over would therefore likely have no benefit, but would certainly incur a small performance penalty.

The non-functional close() method is removed at the same time. It would only make sense if we offered success() and failure(), which we don't.