Improve error message for run() with unblessed reference
Queries that are not plain (scalar) strings only make sense if they can be stringified to valid Cypher. Blessed references can do this by overloading "". Since this isn't possible for unblessed references, passing those instead of a query will always fail.

This is usually a coding error. In this case, the best course of action is for this module to croak its own error message, rather than rely on the Perl runtime.