Rename Protocol to Transport
The classes in the Neo4j::Driver::Protocol namespace actually have little to do with protocols at all. It's just that their main difference is the network protocol being used to transport data to and from the server.

Actually these are adapter classes which hande the details of that data transport. I'm not entirely sure Transport is the right namespace, but it's certainly better than Protocol.