Obtain query statistics by default
So far, the result summary containing query statistics was always available by default when using the Bolt transport, but had to be explicitly enabled manually when using HTTP. This kind of subtle difference complicates API use considerably.

That the more modern Bolt transport always provides query statistics suggests there is no measureable performance hit on the server at all. Requesting and transporting the statistics over HTTP probably does have an influence on performance, but it is believed to be negligible.

At any rate, the advantages that a consistent API provides both for clients and for testing probably outweigh any performance hit. For use cases where HTTP performance is important, it is for now still possible to manually disable the statistics. But this feature may be removed in future if it can't be shown that there in fact is any performance impact.