Implement database simulator
To avoid the requirement of a live connection to a Neo4j server, this commit adds a module that stores canned Neo4j responses that can be used during testing.

This commit also includes reworking of some tests as far as required to support the canned responses, as well as some test optimisations and minor fixes.