Add experimental record streaming interface
This is a first rough implementation for testing purposes. In order to test the upcoming Neo4j::Bolt::ResultStream support, a stream is simulated even for JSON reponses received via HTTP. This is not necessarily particularly efficient. JSON responses are already fully complete and should simply be added directly to the buffer, making the result detached immediately.