Restrict fragile tests for internals to author testing runs
Some tests look into undocumented behaviour, because in certain cases, if such behaviour changes, it should not be an unintentional change. However, because undocumented implementation details may change at any time, these tests tend to be fragile. Therefore, such tests should not be a part of the main test suite, which the CPAN tool runs by default on installation. So these tests make more sense as extra author tests.

This commit also removes one of those fragile tests, which was failing since the previous commit. That test tried to simulate bogus data received from the REST server by meddling with the internals of the StatementResult. However, due to internal implementation changes to StatementResult, this kind of meddling is at present no longer possible.