Attach record streaming interface to Neo4j::Bolt adapter
Up until now, the Bolt result stream was always detached immediately to avoid issues with prematurely closed sessions. This commit finally enables keeping Bolt result streams attached to a live Bolt connection.

Perl uses reference counting to determine when to call the DESTROY method of objects. Since Neo4j::Bolt::Cxn closes the session in its destructor, it is important that our StatementResult holds a reference to the Neo4j::Bolt::Cxn instance until we are certain that we won't need the stream any more.

If the reference is released prematurely, difficult-to-debug issues may arise: Perl does not guarantee to call DESTROY at any particular time, so the code may break in some circumstances but not in others.