Document methods returning lists to expect list context
This primarily applies to the following three methods, which previously were documented to return an arrayref, with calling in list context allowed as experimental feature. Their behaviour is unchanged for backwards compatibility, but calling in scalar context is now experimental for keys() and notifications(); it may be changed in future:
StatementResult->keys()
StatementResult->list()
ResultSummary->notifications()

Additionally, the following three methods have only recently been introduced and not yet released to CPAN. Their behaviour is changed to croak when called in scalar context. This, too, may be changed in future:
Type::Node->labels()
Type::Path->nodes()
Type::Path->relationships()

The run() method offerend by Session and Transaction is not affected, though it will likely be deprecated soon.