Get ServerInfo automatically when opening a session
Allows accessing ServerInfo from the ResultSummary in addition to the Session. Also improves the alignment of the behaviour of HTTP and Bolt connections with each other.

On HTTP connections, the ServerInfo is obtained through Neo4j's Discovery API, which also provides the correct REST endpoints for the particular server. This change will make it easier to address issue #6.

The session config option that allows connecting to a named database is now ignored for older Neo4j server versions that don't support multiple databases.