Move TODO file install location to Driver namespace
The TODO.pod file is mostly a development aid and should be in the root of the repository for easy access. However, once the distribution is packaged and installed, files in that location will be installed right at a top-level namespace inside the local site_perl directory.

The solution I found is basically to copy the file into the Driver namespace before the build, then remove it back afterwards. Because the PodWeaver then starts doing its thing, I copy the original file once again afterwards.

This is a somewhat hacky approach and I can't help thinking that there must be a better way.