Warn if clients access internals of Neo4j structural types
Changes nodes, relationships, and paths to overload the direct hash or array access operator so that a deprecation warning is issued upon use.

Existing code that directly accesses the property hash or path array using ->{} or ->[] should continue to work as before for now (although a warning is issued). Existing code that uses the documented object methods to access properties and path elements will not be affected by this change.

Nodes and relationships are now implemented as a blessed scalar reference. Paths are now implemented as a blessed hash reference. These internal details are subject to change in future.

The experimental Cypher type system customisation feature now offers a new _private() method to access a data store specific to the individual object. This new method should avoid any need to use the old-style access syntax and works even for paths, for which such private data previously wasn't supported.

Resolves #8.