Add support for Neo4j 1.x ServerInfo discovery
Statements still cannot be run because Neo4j 1.x lacked the Transactional HTTP API.

It should actually be possible to map autocommit transactions onto the old cypher endpoint. Such a change would likely be simple, but I really don't want to spend any time on testing it. It's just too old to be worth it.