Rename StatementResult to Result
The Neo4j Driver API changed the name StatementResult to Result, presumably because it's shorter. In Neo4j::Driver, result objects may not always represent statements that have actually been run on the server, so the more general name Result seems more appropriate anyhow.

While this module is part of the public API, the module name is used only internally. A new() method is not publicly exposed. Therefore no issues due to this change are expected.