Change documentation of return values to Neo4j::Types
There shouldn't be many problems with this change, because the behaviour of all methods will remain the same. However, code that checks for the type of a return value using ref() or isa() will eventually need to be adapted.

The warning about pre-0.13 syntax is now superfluous, as 0.18 has finally introcuded deprecation warnings for it.