Introduce http_reason and result_handlers methods
The reason phrase is only relevant in error conditions, which are hopefully rare. By removing it from http_header(), we can slightly reduce the cost of that method.

The result_handlers() method opens up the driver for custom result handlers. One example for how this might be used is a fork of perlbolt that provides results in exactly the format required for the driver, so that no further result handling is required.