Add LWP networking adapter
Introduces a new net_module that uses LWP directly instead of going through REST::Client. However, the driver still depends on REST::Client for testing, which at this point is treated as a runtime requirement by AutoPrereqs.

A consequence of this change is that LWP::Protocol::https, which has been a prerequisite for REST::Client, might no longer always be available, causing HTTPS connections to fail. The documentation is updated to reflect this.