Convert simulator into a custom networking module
REST::Client is no longer used, so the dependency on it can be dropped. However, the ...Net::HTTP::REST adapter still works fine, so there's no particular reason to remove it from the repository (for now).

This change also introduces the TEST_NEO4J_NETMODULE env var, which can be used to run the driver's test suite against a custom net_module. This allows streamlining the simulator's store() operation.