Update driver for Neo4j 4.2.5
Neo4j version 4.2.5 changed the 4.2.0 Jolt media type to conform to RFC 7464. This is a breaking change for users of Jolt. Jolt has been experimental and disabled by default in this driver in part because such a change was already anticipated; see neo4j/neo4j#12663.

This commit addresses the problem by treating both RS and LF as Jolt event separators while ignoring any empty events. This is an interim solution pending optimisation of the Jolt event parsing; see #9.