Change experimental "run multi statements" to explicit method
The earlier approach to change the behaviour of run() based on its context and attribute type can be needlessly confusing in the driver's implementation as well as the user's code. It is now deprecated. It's much better to offer a dedicated method to run multiple statements at once.

In the long run, the plan is still to use lazy execution for this feature. Jolt makes that worthwhile even in HTTP mode. However, this optimisation is not a priority right now. Once it does land, there shouldn't be a need for _run_multiple() anymore. Plus, even right now, it's a bit of an edge case to even include support for running multiple statements at all. That's why the new method is named like a private method.

Adding Try::Tiny to @CARP_NOT ensures the deprecation warning is given in the correct scope for Bolt autocommit transactions.