Change scalar context return of list methods to return item count
Methods that usually return lists can also be called in scalar context. The best practices for the scalar return value for such cases in Perl are somewhat unclear. <https://www.perlmonks.org/?node_id=311537> lists several options and has some discussion around that.

The least surprising option appears to be to mimic the behaviour of a list variable when evaluated in scalar context, returning the count of list items. Even novice users of Perl should be familiar with that. Plus, the rest of the language is designed around plain lists, so combining these methods with built-ins such as grep or map works very well like that.

Another option that has been considered is the use of array references, but those are problematic for empty lists because [] is truthy and undef tends to throw warnings. It is also a bit tricky to obtain the number of items in the list that way. So this seems a bit more difficult to handle than when just returning the item count.

Prior to version 0.13, the driver documentation suggested that array references should be expected by default for ResultSummary->notifications() and StatementResult->keys(). However, prior to 0.13, the documentation also emphasized that the interface was not yet stable, and the change was documented in the changelog for 0.13, which was released almost two years ago. Since then, the scalar return value has been listed as experimental. So, this change may break some really old and unmaintained code, but it should hopefully not be too big of a surprise for anyone.

The Result->list() method's behaviour is unchanged.