Rename config option tls_ca to trust_ca
With the tls option gone, the name tls_ca doesn't make much sense anymore. However, the Neo4j Driver API does not have a common name for this option. Every one of the five official drivers uses a different name. A recurring theme among the official drivers is that this particular option is part of the "trust strategy" to be used, therefore trust_ca seems like a reasonable choice.