Cache ServerInfo in driver object
Neo4j sessions are supposed to be lightweight. To reduce session creation overhead, this change introduces caching of the ServerInfo discovery results in the driver object, so that the extra roundtrips are only required for the first session. The default database name gets the same treatment and is stored in the SessionInfo object as well.

For HTTP, creating additional sessions will now not cause any network traffic anymore. The performance gain depends on how many sessions you create, and how high your network latency is. A gain is hardly measurable when running the driver test suite locally on my development machine, but YMMV.

For Bolt, however, every session will still create a new libneo4j-client connection, which means running through the Bolt Handshake Protocol and sending a HELLO message. Because libneo4j-client completely hides server sessions and state from its clients, this is currently unavoidable.