Tag methods without pod documentation as uncoverable
The documentation of deprecated methods is not included in the module's pod. It lives in the additional document Deprecations.pod. This is done in order to keep the module pod as concise and easy to understand as possible.

For most of the driver packages, the new() method is only for internal use. Clients themselves only use new() to create a Neo4j::Driver object; all other objects will be created by the driver. This effectively makes new() in those other modules a method that is meant to be private. However, the pattern of creating an object using new() is well-established enough that changing the name to _new() to reflect this seems ill-advised. At the same time, pod documentation is not appropriate for a private method.

Experimental methods are already well-documented in pod, but Devel::Cover doesn't pick these up automatically.

Explicitly declaring all these cases as "uncoverable" helps to discover subs that are actually undocumented.