Add nested_tx config option and associated warnings
This driver's API should be uniform for Bolt and HTTP. Since Bolt doesn't support nested transactions, these should in principle be disallowed for HTTP as well. However, for backwards compatibility, we'll start by giving warnings.

On HTTP connections, users should have a way to enable nested transactions if needed. This change introduces a new config option for this purpose. For now, this option is enabled by default, but the intention is to change this soon.

The warnings can only be given on a best-effort basis due to the design of the driver's HTTP controller. Ideally, warnings were given when beginning additional transactions; instead, warnings are only given when a second transaction is actually used (e.g. with run or commit). This may or may not change in future.

Once neo4j: URIs are implemented, they should not allow nested transactions. However, due to that same design issue, we may have to be content with best-effort warnings for neo4j: as well. Those warnings should indicate that they may become fatal. Since neo4j: is intended to become the default URI scheme once implemented, that indication is present right from the start.