Add initial plug-ins implementation
The design problem with event handlers is that events differ by the results they yield in the presence of multiple handlers. In some cases, you would need exactly one result value (e. g. the adapter factory). In other cases, you'd want to gather a variety of values in a list (e. g. the old result_handler method). For some events, handlers might even need a way to control how other handlers are invoked in order to avoid side-effects.

The solution chosen here offers a maximum of flexibility to event handlers. It does so by providing a callback to each handler that is used to invoke the next handler for that same event. Plug-ins can choose to invoke this callback in the beginning or at the end of handlers, they can invoke it conditionally, or they can choose to not invoke it at all. However, this also means that simply forgetting to invoke the callback stops other plug-ins from ever receiving that event.

This will probably not be a big problem because for a software with as limited a scope as the driver, in all likelyhood the number of loaded plug-ins is small to begin with. The number of cases where one event has multiple handlers defined for it will likely be equally small. So, in practice, how (or if) that callback is invoked usually won't really make a difference.

That said, this also means that this solution could be simplified for greater ease of use, at the cost of possibly breaking backwards compatibility. It seems worthwhile to consider revisiting this later. The plug-in API should probably keep the label "experimental" at least until version 1.00 or thereabouts (perhaps even longer).