Recommend against the use of sub signatures in plug-ins
For event handlers, the caller defines the API and the subroutine needs to match it. This is the inverse of the usual procedural model, where the subroutine (the callee) defines the API to be matched by the caller.

When additional arguments are "optional", the usual procedural model takes this to mean it's optional to pass them into a sub call. In the context of an event handler, these arguments are always passed, but may optionally be ignored by the sub.

That's a problem in the context of Perl's subroutine signatures, which have strict arity checking. It can make it difficult or impossible to add additional arguments to existing event handlers in future.

This change recommends that plug-in authors don't use sub signatures at all. It does so by briefly explaining the issue and by removing subroutine signatures from all verbatim code snippets.

Several of the code snippets become more verbose and less easy to understand at a glance. This is unfortunate, but given the potential for hidden bugs here, it seems to be the lesser evil.