Move net adapter API docs from Net to Plugin
Note that the new() method is not required for an HTTP adapter because the plug-in factory event handler will be in charge of instantiating the object. It may do so using new(), but it may also choose another option. However, for the soon-to-be deprecated net_module option, new() is still required.