Remove POD from Result implementations
For the version 0.x driver series, these classes are an implementation detail that clients are generally supposed to ignore.

Eventually it may be necessary to formally specify the interface to each result variant's internals for the result handler API (for use by plug-ins). But this has been postponed until some time after the 1.00 release.