Fix dependency chain in 5.16 test environment
It seems like the shogo82148/actions-setup-perl GH action, which pre-installs or updates a number of toolchain modules and other CPAN distros, creates an incompatibility with the dependency chain of the driver when run on Perl v5.16. The error message says:

! Installing the dependencies failed: Installed version (6.63_02) of ExtUtils::MakeMaker is not in range '6.78'
! Bailing out the installation for Encode-3.19.

However, Encode-3.19 doesn't actually seem to have a dependency on ExtUtils-MakeMaker-6.78, so I'm not sure what's going on there. I do note that the version of HTTP-Message that cpanm wants to install has a dependency on Encode-3.01, whereas a Perl v5.16-era contemporary HTTP-Message would have been content with Encode-2.12.

*sigh* I think it may be time to remove LWP from the dependency chain entirely.

Alternatively, take steps to ensure that the Perl 5.16 testing environment offers a known-good set of contemporary module versions. I mean, testing with the newest module versions on v5.16 doesn't seem that useful anyway.