Rework to use Neo4j::Error for error handling, add error event
All Neo4j server errors and network errors are now triggered using the "error" event handler, which plug-ins can choose to add custom handlers to. These errors are reported as Neo4j::Error instances, which gives plug-ins the option to query structured error metadata using an OO interface. It also gives the driver the opportunity to make additional details about the error condition available to clients, such as the raw server response in case of decoding errors. Resolves #7.

The default handler for the "error" event tries to recreate the old error message strings for backwards compatibility. However, the order of messages had to be changed in some cases, with the most significant error now always being in first place. Additionally, Neo4j server errors received over a Bolt connection are now reported on a single line for consistency with HTTP and with other error types.