Remove Perl core modules from dependency list
Overly long dependency lists can be difficult to parse for humans, hence we hide stuff that really is irrelevant. Ideally, a Dist::Zilla plug-in would do this automatically based on Module::CoreList, so that humans don't forget it.

The downside of this strategy is a potential snafu in the (somewhat unlikely) event that a module is removed from Perl core and people try to use an old version of this software with a new Perl. Therefore, a Dist::Zilla plug-in should probably add such dependencies at least as "suggested" in some way.

Then again, an argument could be made that it is really MetaCPAN that should learn how to properly reduce dependency lists by combining the declared requirements of modules with the declared minimum version of Perl.