Reduce risk that execute tests trigger race condition
While it's encouraging to see that most smokers are fast enough to succeed on the time-based retry test even with the previously very tight timeout, it seems sensible to adapt the timing of this test to the capabilities of the test system. This particular solution is probably overengineered, but hey it works.

Prints a measure of the determined machine speed as diagnostic output during AUTOMATED_TESTING. This probably should have been included previously.