Fix bug in reporting Neo4j::Bolt::ResultStream errors
During the error reporting rework in 0.36, I blundered the invocation of _trigger_bolt_error() in Neo4j::Driver::Result::Bolt. The test suite didn't cover that particular code path.

In practical usage, this bug is only triggered rarely, because most of the time, Neo4j errors are raised at the time when the query is run on the connection/transaction, before the result stream is ever used. This kind of error would be raised already during the call to run(). Errors raised during run() were triggered correctly even in 0.36.