Move driver config into a hash of its own
Clearer separation of concerns and easier passing of driver configuration to submodules and closures. Lays the groundwork for fixing the memory leak in the http_adapter_factory default event handler, and also for event handlers in general having access to driver config options.

Attempts by users to set options directly in the driver hash rather than through new() or config() will continue to work, but only for a limited period of time. A warning will be issued.