Add support for spatial and temporal Cypher values
This change adds simple regex-based parsing of ISO temporal values and pseudo-WKT spatial values. To avoid unneccessary memory allocation, the existing Jolt hash is reused. Regex parsing happens lazily with the result being cached in the hash, so that merely retrieving values as well as subsequent method calls are inexpensive.

MockQuery is extended to allow for the use of arbitrary Jolt values as results. This simplifies testing of temporal values.

Temporal values won't work with Neo4j::Types v1 because it lacks the required packages. A future driver version will require v2. For now, we load those packages at runtime when required (with require instead of use) so that the rest of the driver continues to work with v1.