Add basic implementation of neo4j URI scheme
Neo4j defines neo4j: URIs as being equivalent to bolt: for single servers. However, we can't assume the user will have Neo4j::Bolt installed. Therefore, this implementation naively checks in %INC whether Neo4j::Bolt is already loaded; if so, it will treat neo4j: URIs as bolt: URIs, otherwise as http: URIs. As a result, simple neo4j: URLs should work correctly "out of the box" most of the time.

Since neo4j: URIs are treated as identical to http: most of the time, switching the default scheme to neo4j: at this time should be fairly painless.

A future version might improve neo4j: handling such that the driver actually tries a Bolt connection and uses it if successful. However, a potential snag here is that in spite of best efforts, Bolt and HTTP don't quite behave identically. There's a small risk of applications breaking because the protocol changed unexpectedly. It seems sensible to recommend against the use of neo4j: URIs with this driver until this issue can be addressed.

Client-side routing to enterprise clusters is out of scope for this driver. I'm open to adding a plugin hook for it in the future, but it's not a big priority.