Make consume() stable, discourage summary()
Change consume() to actually exhaust the result stream, rather than just setting the exhausted flag. This is potentially important, because the summary may otherwise not be reliably available, depending on network protocol.

Calling fetch() in a loop is assumed to be better than using _fill_buffer(), because fetch() releases the memory used for each record right away. This saves considerable amounts of memory for large results, especially when streaming is used.

There is significant potential for optimising consume(), but all of the options are either not worthwhile or not feasible right now. It's useful to remember that in practice, consume() is almost always called when the result stream is already exhausted anyway.

Bolt: Calling neo4j_update_counts() directly might be an option, but perlbolt actually requires us to fetch everything first anyway. So this would require changes to perlbolt. The one thing we could do here is to skip deep_bless, but that'll be removed soon anyway.

Jolt: We could iterate to the end of the raw Jolt event stream, dropping everything without ever decoding any JSON until we reach the summary. However, with Jolt async fetching not publicly implemented and with Jolt results still relying on gather_results, it's not yet the time to optimize consume().

JSON: The stats are always available. But JSON is almost exclusively used in legacy setups, so it isn't worth spending time on.

The official drivers have largely removed summary() a while ago. This driver should follow suit. The 1.00 update seems like a good time to deprecate summary().