Avoid retry test failures
The sleep function can be a bit unreliable, in rare cases waiting much longer than requested. This effect can prevent expected retries from being attempted at all because of the timeout.

Logic that depends on a certain amount of retries taking place is inherently faulty, but is still required for this particular test. The sensible course of action is to limit this check to author testing. Tests are set up to fail on Perl warnings during author testing only, so switching from "ok" to "warn" in this change achieves just that.