Work around crash of types-temporal.t on Win32
The issue here is Win32 not allowing dates before 1970 in strftime (Cygwin is not affected). Given that perl does support such dates, and that Time::Piece is a core module, this is arguable an issue with perl itself.

Because strftime is only used in the test, the Neo4j::Driver::Type::DateTime module is expected to work correctly in Win32.

This change simply skips the problematic test on Win32. An alternative might be to use strftime from DateTime instead, which I believe to be implemented in pure Perl, so it'd probably work fine. However, I don't want to require DateTime on every OS just because of the test.