Relax SV type strictness of get() args in Record
When the issue of index/key ambiguity was fixed in Record->get() years ago, I overlooked that scalars like `0` and `"0"` should generally be treated identically in Perl: For a query like `RETURN 1 AS foo`, `get("0")` should succeed, with the argument treated as numeric index 0.

This change ensures that this is in fact the case. To allow for fast lookups, all unambiguous indices and keys are hashed in ResultColumns, with any potentially ambiguous index/key moved into a separate hash.

We currently have plenty of space in the Record hash ref to store that extra data. However, one of the anticipated efficiency gains 1.00 will enable is to simplify these data structures. Anticipating that, these two hashes are nested, with the inner hash using the outer hash's empty string entry. Perl syntax allows the empty string in this context, Neo4j syntax does not, so this works fine.