Allow cypher_params 'v2'
The literal v-string is easy to express in Perl, but if you import the config from YAML or something, you may run into trouble. Plus, even in Perl, it's kind of an obscure feature that might confuse some users. The best solution seems to be to allow both the quoted and the unquoted form.