Rewrite synopsis sections
For most modules, the new synopsis is written as a summary of the most important methods and ways to use the module. This is hopefully more immediately useful for the majority of people than the example code used as synopsis prior to this change.

I want to include complete working examples in this distribtion as separate files eventually, but it's currently not a priority.

The example-style synopsis is retained for Neo4j::Driver itself. I believe this is useful for the main module as an exception, as it shows some of the most commonly used classes and methods in the entire distribution at a glance.

In addition to the synopsis, this change adjusts parts of the module descriptions to try and make a more coherent whole.