Update docs for bugs and limitations
With this software moving towards the 1.00 release, some of the documented issues are to be considered minor caveats rather than significant limitations. In general, the driver by now is proven enough to make blanket notices to "look out for new bugs" etc. unnecessary. That issue reports are welcome should go without saying for any software, and the same goes for further optimisation.

The Neo4j::Driver::Net document is also updated to the current version, in particular with regards to the described limitations.