Fix docs: Add undeclared concurrent tx to deprecation list
Opening concurrent transactions without first enabling them using the config option is an experimental feature that was deprecated in 0.39 (or 0.30, depending on how you look at it).

Because this is technically just the software enforcing a specific way to use it, rather than an outright feature removal, the concept of deprecation wasn't mentioned at the time. But the language used was very clear from the beginning that the old way of doing things would become a fatal error eventually. Stark Perl warnings are issued by the driver when trying to use it. So, in practice, it was a deprecation, even if that word wasn't actually used.

For the upcoming 1.00 release changelog, it's useful to treat concurrent transactions without first enabling them exactly the same as any other deprecation. This change adds this item to the list of deprecations.