Discourage Result->list() in scalar context
All other methods in the driver and its types returning arrays yield the size of the array in scalar context. This is the only exception, and there isn't a real good reason for it, other than "it's always been that way".

It seems that changing the scalar context would be ill-advised because of existing users. Nevertheless, it's probably a good idea to very gently nudge users in the right direction.