Remove overload from byte array object stringify
Jolt byte arrays were originally returned as a regular unblessed string to match old versions of Neo4j::Bolt. Neo4j::Types v2 defined an object type for byte arrays, implemented in the driver since version 0.46.

Deprecation warnings for direct string access of byte value objects were introduced at the same time, using overload. While we could continue to overload stringify to keep the deprecated old behaviour working, using overload has a memory overhead, so it'd be good to remove it eventually. 1.00-TRIAL throws an exception on stringify, which is not a great user experience. But just removing overload entirely has a risk of things breaking silently.

To avoid the latter, it seems sensible to continue to use overload to keep issuing a warning for maybe a full year or so.