Exclude Deprecations.pod and Net.pod from dist
The discussion of deprecated 0.x features is only of historical significance and can be removed entirely.

The discussion of the driver's networking modules is still useful, but probably only for those who want to work on the driver itself, not so much for those who merely want to use the driver. So Net.pod doesn't need to be on CPAN.

Closes #19.