Drop Bolt deep_bless
The minimum supported version of Neo4j::Bolt is now 0.4201, which is the first version to require Neo4j::Types. This means that converting Neo4j::Bolt values to Neo4j::Driver::Type values by traversing the tree of result values is no longer necessary.

As a result, receiving complex query results (e. g. with many columns, lists, maps, or properties) over Bolt is about twice as fast as before. The overhead of Neo4j::Driver is eliminated almost entirely for such queries.

The performance of Bolt with Neo4j::Driver continues to be limited by the general handling of queries and results, which is unaffected by this change. Very simple queries (e. g. having just a single scalar column) may still take up to twice as long as they would take if ran directly with Neo4j::Bolt. But such simple queries are typically very fast anyway, so this speed difference may not be too important for many applications.