Refactor structural type packages to match Jolt
With this change, Jolt values can be passed through with only minimal changes in deep_bless. As a result, receiving query results from Neo4j 4.3 and later over HTTP is faster than before (between 2 % and 20 % during my tests, with more complex results benefiting more).

The semantics of IDs in Jolt results are determined by the Jolt version (v1: legacy ID, v2: element ID). Separate "V1" packages are introduced to cleanly separate the differences between Jolt versions.

I rejected three alternative solutions:

(1) Element IDs are always strings, legacy IDs always numbers. Perl v5.36+ has builtins to easily distinguish between those. However, earlier versions would need custom code. While length() could currently be used for that in an efficient way (element IDs >= 40, legacy IDs <= 20), element IDs are expressly defined to be opaque and their format may change in future. So we'd still need to run looks_like_number heuristics for short IDs, which is comparatively slow and a bit messy.

(2) Since the Jolt version is known when blessing the node/rel value, it would be easy to mark legacy IDs in a special, non-destructive way: by making them a reference. However, even though that would likely be fast and effective, I think it'd be confusing to read; at least as messy as the heuristics.

(3) It would be easy to push a version marker onto the array's end when blessing the node/rel value. Because of the default value for PERL_ARRAY_NEW_MIN_KEY, there is even a spare slot in the node array, so this would be really cheap. Relationships are larger, and pushing onto them requires a Newx+Copy in av_extend(), which feels like a bit of a waste; that said, the performance impact of this seems to be neglegible in my tests.

All three alternatives require branches in the id() methods, which makes them more complex and thus inferior to the chosen solution.